bust out KeyScan into its own file for cleanliness
